// packages/terrain/infrastructure/primitives.d.ts
import { NoiseServicePort } from '../domain/noise-service-port';
import { type NoiseFn2D, type NoiseFn3D, type RandFn } from './perlin';
export declare const mulberry32: (seed: number) => RandFn;
export declare const normalizeNoise: (value: number) => number;
export declare const WEYL_C = 2654435761;
export declare const WEYL_E = 3144134277;
export declare const WEYL_W = 1013904242;
export declare const WEYL_J = 2773480762;
export declare const WEYL_3D = 2654435769;
export declare const SCALE_C = 0.0005;
export declare const SCALE_E = 0.001;
export declare const SCALE_W = 0.002;
export declare const SCALE_J = 0.02;
export declare const toPV: (w: number) => number;
export declare const computeOctaveNoise: (noiseFn: NoiseFn2D, x: number, z: number, octaves: number, persistence: number, lacunarity: number) => number;
export type TerrainChannelSamples = Readonly<{
    continentalness: Float64Array;
    erosion: Float64Array;
    pv: Float64Array;
    jaggedness: Float64Array;
}>;
export declare const computeTerrainChannels: (noiseFnContinentalness: NoiseFn2D, noiseFnErosion: NoiseFn2D, noiseFnWeirdness: NoiseFn2D, noiseFnJaggedness: NoiseFn2D, xStart: number, zStart: number) => TerrainChannelSamples;
export type NoisePrimitives = Readonly<{
    raw2D: NoiseFn2D;
    raw3D: NoiseFn3D;
    continentalness: NoiseFn2D;
    erosion: NoiseFn2D;
    weirdness: NoiseFn2D;
    jaggedness: NoiseFn2D;
    noise2D: (x: number, z: number) => number;
    octaveNoise2D: (x: number, z: number, octaves: number, persistence: number, lacunarity: number) => number;
    noise3D: (x: number, y: number, z: number) => number;
    continentalnessAt: (x: number, z: number) => number;
    erosionAt: (x: number, z: number) => number;
    weirdnessAt: (x: number, z: number) => number;
    jaggednessAt: (x: number, z: number) => number;
    sampleTerrainChannels: (xStart: number, zStart: number) => TerrainChannelSamples;
}>;
export declare const createNoisePrimitives: (seed: number) => NoisePrimitives;
export declare const noise2DBatchXY: (primitives: NoisePrimitives, xs: ReadonlyArray<number>, zs: ReadonlyArray<number>) => ReadonlyArray<number>;
export declare const noise3DBatchXYZ: (primitives: NoisePrimitives, xs: ReadonlyArray<number>, ys: ReadonlyArray<number>, zs: ReadonlyArray<number>) => ReadonlyArray<number>;
export declare const octaveNoise2DBatchXY: (primitives: NoisePrimitives, xs: ReadonlyArray<number>, zs: ReadonlyArray<number>, octaves: number, persistence: number, lacunarity: number) => ReadonlyArray<number>;
export declare const noise2DBatch: (primitives: NoisePrimitives, points: ReadonlyArray<readonly [number, number]>) => ReadonlyArray<number>;
export declare const octaveNoise2DBatch: (primitives: NoisePrimitives, points: ReadonlyArray<readonly [number, number]>, octaves: number, persistence: number, lacunarity: number) => ReadonlyArray<number>;
export declare const buildNoisePortFromPrimitives: (primitives: NoisePrimitives, seed: number) => NoiseServicePort;
//# sourceMappingURL=primitives.d.ts.map